+++ apps/back-end/services/shop/src/shop/entities/shop.entity.ts
@@ -1,11 +1,11 @@
-import { ObjectType, Field, Int, ID } from '@nestjs/graphql';
+import { ObjectType, Field, Int, ID, Float } from '@nestjs/graphql';
 import { Product } from './product.entity';
 @ObjectType()
 export class Location {
-  @Field((type) => Int)
+  @Field((type) => Float)
   lat: number;
 
-  @Field((type) => Int)
+  @Field((type) => Float)
   long: number;
 
   @Field((type) => String)
